# api-docs.json
{"openapi":"3.1.0","info":{"title":"메뉴 관리 API 명세서","description":"Spring Data JPA 수업용 프로젝트(chap06)의 API 명세서다. 메뉴와 카테고리 조회, 메뉴 등록·수정·삭제를 제공한다.","contact":{"name":"Ohgiraffers 강사팀","email":"owl.ohgiraffers@gmail.com"},"version":"v1"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Menu","description":"메뉴 조회 및 관리 API"},{"name":"Category","description":"카테고리 조회 API"}],"paths":{"/api/menus/{menuCode}":{"get":{"tags":["Menu"],"summary":"메뉴 상세 조회","description":"정상 응답은 ResponseMessage 형식이며 result.menu에 조회한 메뉴가 담깁니다.","operationId":"findMenuByCode","parameters":[{"name":"menuCode","in":"path","description":"조회할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"메뉴 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"해당 코드의 메뉴가 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"put":{"tags":["Menu"],"summary":"메뉴 수정","description":"정상 응답은 ResponseMessage 형식이며 result.menu에 수정된 메뉴가 담깁니다.","operationId":"updateMenu","parameters":[{"name":"menuCode","in":"path","description":"수정할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuDTO"}}},"required":true},"responses":{"200":{"description":"메뉴 수정 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"요청 값 또는 카테고리 코드가 잘못됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"수정할 메뉴가 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"delete":{"tags":["Menu"],"summary":"메뉴 삭제","description":"정상 응답은 ResponseMessage 형식이며 result.deletedMenuCode에 삭제된 메뉴 코드가 담깁니다. 실제 HTTP 상태 코드는 200이고 ResponseMessage의 httpStatus 값은 204입니다.","operationId":"deleteMenu","parameters":[{"name":"menuCode","in":"path","description":"삭제할 메뉴 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"메뉴 삭제 성공(응답 본문의 httpStatus는 204)","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"삭제할 메뉴가 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus":{"get":{"tags":["Menu"],"summary":"전체 메뉴 조회","description":"정상 응답은 ResponseMessage 형식이며 result.menus에 메뉴 목록이 담깁니다.","operationId":"findAllMenus","responses":{"200":{"description":"메뉴 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}},"post":{"tags":["Menu"],"summary":"메뉴 등록","description":"정상 응답은 ResponseMessage 형식이며 result.menu에 등록된 메뉴가 담깁니다.","operationId":"saveMenu","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MenuDTO"}}},"required":true},"responses":{"201":{"description":"메뉴 등록 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"요청 값 또는 카테고리 코드가 잘못됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/search":{"get":{"tags":["Menu"],"summary":"가격 기준 메뉴 검색","description":"정상 응답은 ResponseMessage 형식이며 result.menus에 기준 가격을 초과하는 메뉴 목록, result.searchPrice에 검색 기준 가격이 담깁니다.","operationId":"findMenusByPrice","parameters":[{"name":"menuPrice","in":"query","description":"검색 기준 가격. 이 가격을 초과하는 메뉴를 조회합니다.","required":true,"schema":{"type":"integer","format":"int32"},"example":10000}],"responses":{"200":{"description":"메뉴 검색 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"가격 파라미터가 없거나 형식이 잘못됨","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/pages":{"get":{"tags":["Menu"],"summary":"메뉴 페이지 조회","description":"페이지 번호는 1부터 시작합니다. 정상 응답은 ResponseMessage 형식이며 result에 content(메뉴 목록), totalElements(전체 항목 수), totalPages(전체 페이지 수), size(페이지 크기), number(현재 페이지 번호), first(첫 페이지 여부), last(마지막 페이지 여부)가 담깁니다.","operationId":"findMenuPage","parameters":[{"name":"pageable","in":"query","description":"페이지 정보. page는 1부터 시작하며 size와 sort를 함께 지정할 수 있습니다.","required":true,"schema":{"$ref":"#/components/schemas/Pageable"},"example":1}],"responses":{"200":{"description":"메뉴 페이지 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"잘못된 페이지 요청 값","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/menus/pages/sort":{"get":{"tags":["Menu"],"summary":"정렬된 메뉴 페이지 조회","description":"페이지 번호는 1부터 시작합니다. 정상 응답은 ResponseMessage 형식이며 result에 content(메뉴 목록), totalElements(전체 항목 수), totalPages(전체 페이지 수), size(페이지 크기), number(현재 페이지 번호), first(첫 페이지 여부), last(마지막 페이지 여부), sort(정렬 필드), direction(정렬 방향)이 담깁니다.","operationId":"findMenuPageWithSort","parameters":[{"name":"pageable","in":"query","description":"페이지 정보. page는 1부터 시작하며 기본 size는 5입니다.","required":true,"schema":{"$ref":"#/components/schemas/Pageable"},"example":1},{"name":"sortBy","in":"query","description":"정렬 기준 필드","required":false,"schema":{"type":"string","default":"menuPrice"},"example":"menuPrice"},{"name":"direction","in":"query","description":"정렬 방향(asc 또는 desc)","required":false,"schema":{"type":"string","default":"asc"},"example":"asc"}],"responses":{"200":{"description":"정렬된 메뉴 페이지 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"400":{"description":"잘못된 페이지 또는 정렬 요청 값","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/categories":{"get":{"tags":["Category"],"summary":"전체 카테고리 조회","description":"정상 응답은 ResponseMessage 형식이며 result.categories에 카테고리 목록이 담깁니다.","operationId":"findAllCategories","responses":{"200":{"description":"카테고리 목록 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/categories/{categoryCode}":{"get":{"tags":["Category"],"summary":"카테고리 상세 조회","description":"정상 응답은 ResponseMessage 형식이며 result.category에 조회한 카테고리가 담깁니다.","operationId":"findCategoryByCode","parameters":[{"name":"categoryCode","in":"path","description":"조회할 카테고리 코드","required":true,"schema":{"type":"integer","format":"int32"},"example":1}],"responses":{"200":{"description":"카테고리 조회 성공","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"404":{"description":"해당 코드의 카테고리가 없음","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}},"500":{"description":"서버 내부 오류","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}}},"components":{"schemas":{"MenuDTO":{"type":"object","description":"등록할 메뉴 정보","properties":{"menuCode":{"type":"integer","format":"int32"},"menuName":{"type":"string"},"menuPrice":{"type":"integer","format":"int32"},"categoryCode":{"type":"integer","format":"int32"},"categoryName":{"type":"string"},"orderableStatus":{"type":"string"}}},"ResponseMessage":{"type":"object","properties":{"httpStatus":{"type":"integer","format":"int32"},"message":{"type":"string"},"result":{"type":"object","additionalProperties":{}}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}}}}}